Add metrics and revamp logic (#206)
* Add metrics and revamp logic

Declare prometheus metrics in gossipsub

Add missing metric labels

Fix metrics types

Add onReportValidation

* Compile tests

* Fix tests

* Customize await policy

* Run prettier on files

* Relax heartbeat test condition

* Re-add util package for browser testing

* Remove unused dependencies

* Add getMeshPeers

* Fix rpc metric name

* Improve metrics

* More gossip promise metrics

* Fix metrics typo

* Clarify metrics names

* Track metric of score deltas

* Fix score metrics

* Add buckets for gossipsub_score_cache_delta

* Review PeerScore logic

* Dump peer stats

* Fix PeerStats constructor

* Enable strict typescript checks

* Fix compute-score logic

* f

* Fix peer stats types

* Reenable go-gossipsub tests (#201)

* Add checkReceivedSubscriptions

* Add checkReceivedSubscriptions to 'test gossipsub multihops'

* Add checkReceivedSubscriptions to 'test gossipsub tree topology'

* test gossipsub star topology with signed peer records

* Fix 'test gossipsub direct peers'

* Fix 'test gossipsub flood publish'

* Fix 'test gossipsub star topology with signed peer records'

* 'direct peers' test: wait for subscriptions event again

* 'direct peer': await for 2 peer:connect events

* 'direct peers': add missing Promise.all

* Expect topic peers to contain peer id

* Fix test types

* Set as connected in addPeer()

* Prune publishedMessageIds

* Fix markFirstMessageDelivery typo

* Same logic in scoreMetrics

* More metrics for p3 and p7 (#213)

* Add behaviourPenalty metrics

* Add duplicateMsgDelivery metric

* Observe topic and peer in duplicateMsgDelivery topic

* Remove peerId from duplicateMsgDelivery metric

* Use min meshMessageDeliveriesWindow

* Use topic label for duplicateMsgDelivery metric

* Record duplicateMsgDelivery in all cases

* Forward messages to floodsub peers (#214)

* Forward messages to floodsub peers

* Add comments


* Add missing msgId in validateReceivedMessage (#215)

* Fix 'test gossipsub opportunistic grafting'

* Fix tests in browser

* Fix tests suspended issue in browsers

* Fix 'test gossipsub fanout expiry' go-gossipsub test

* GossipsubIWantFollowupTime as a param (#216)

* Increase resolution of delay metrics (#217)

* Fix minMeshMessageDeliveriesWindow (#218)

* Fix minMeshMessageDeliveriesWindow

* Fix gossipsubIWantFollowupTime for metric

* Fix tracer prune()

* Change to maxMeshMessageDeliveriesWindow

* Use maxMeshMessageDeliveriesWindowSec for metric

* Rename gossipsubIWantFollowupMs option, revert/correct tracer.prune() logic

* Reset behaviourPenalty histogram to track current count

* publish(): return number of sent peers

* Add getScore()

* makePrune: update PeerStat so that we don't apply p3 penalty

* Remove redundant this.score.prune() in heartbeat

* Track iasked cache size per heartbeat and remove TODOs

* validateReceivedMessage: check duplicate message first (#223)

+++ ts/tracer.ts
@@ -1,8 +1,6 @@
-import { GossipsubIWantFollowupTime } from './constants'
 import { messageIdToString } from './utils'
-import pubsubErrors = require('libp2p-interfaces/src/pubsub/errors')
-
-const { ERR_INVALID_SIGNATURE, ERR_MISSING_SIGNATURE } = pubsubErrors.codes
+import { MsgIdStr, PeerIdStr, RejectReason } from './types'
+import { Metrics } from './metrics'
 
 /**
  * IWantTracer is an internal tracer that tracks IWANT requests in order to penalize
@@ -17,50 +15,86 @@ export class IWantTracer {
    * Promises to deliver a message
    * Map per message id, per peer, promise expiration time
    */
-  promises = new Map<string, Map<string, number>>()
+  private readonly promises = new Map<MsgIdStr, Map<PeerIdStr, number>>()
+  /**
+   * First request time by msgId. Used for metrics to track expire times.
+   * Necessary to know if peers are actually breaking promises or simply sending them a bit later
+   */
+  private readonly requestMsByMsg = new Map<MsgIdStr, number>()
+  private readonly requestMsByMsgExpire: number
+
+  constructor(private readonly gossipsubIWantFollowupMs: number, private readonly metrics: Metrics | null) {
+    this.requestMsByMsgExpire = 10 * gossipsubIWantFollowupMs
+  }
+
+  get size(): number {
+    return this.promises.size
+  }
+
+  get requestMsByMsgSize(): number {
+    return this.requestMsByMsg.size
+  }
 
   /**
    * Track a promise to deliver a message from a list of msgIds we are requesting
    */
-  addPromise(p: string, msgIds: Uint8Array[]): void {
+  addPromise(from: PeerIdStr, msgIds: Uint8Array[]): void {
     // pick msgId randomly from the list
     const ix = Math.floor(Math.random() * msgIds.length)
     const msgId = msgIds[ix]
     const msgIdStr = messageIdToString(msgId)
 
-    let peers = this.promises.get(msgIdStr)
-    if (!peers) {
-      peers = new Map()
-      this.promises.set(msgIdStr, peers)
+    let expireByPeer = this.promises.get(msgIdStr)
+    if (!expireByPeer) {
+      expireByPeer = new Map()
+      this.promises.set(msgIdStr, expireByPeer)
     }
 
-    if (!peers.has(p)) {
-      peers.set(p, Date.now() + GossipsubIWantFollowupTime)
+    const now = Date.now()
+
+    // If a promise for this message id and peer already exists we don't update the expiry
+    if (!expireByPeer.has(from)) {
+      expireByPeer.set(from, now + this.gossipsubIWantFollowupMs)
+
+      if (this.metrics) {
+        this.metrics.iwantPromiseStarted.inc(1)
+        if (!this.requestMsByMsg.has(msgIdStr)) {
+          this.requestMsByMsg.set(msgIdStr, now)
+        }
+      }
     }
   }
 
   /**
    * Returns the number of broken promises for each peer who didn't follow up on an IWANT request.
+   *
+   * This should be called not too often relative to the expire times, since it iterates over the whole data.
    */
-  getBrokenPromises(): Map<string, number> {
+  getBrokenPromises(): Map<PeerIdStr, number> {
     const now = Date.now()
-    const result = new Map<string, number>()
+    const result = new Map<PeerIdStr, number>()
 
-    this.promises.forEach((peers, msgId) => {
-      peers.forEach((expire, p) => {
+    let brokenPromises = 0
+
+    this.promises.forEach((expireByPeer, msgId) => {
+      expireByPeer.forEach((expire, p) => {
         // the promise has been broken
         if (expire < now) {
           // add 1 to result
           result.set(p, (result.get(p) ?? 0) + 1)
           // delete from tracked promises
-          peers.delete(p)
+          expireByPeer.delete(p)
+          // for metrics
+          brokenPromises++
         }
       })
       // clean up empty promises for a msgId
-      if (!peers.size) {
+      if (!expireByPeer.size) {
         this.promises.delete(msgId)
       }
     })
+
+    this.metrics?.iwantPromiseBroken.inc(brokenPromises)
 
     return result
   }
@@ -68,18 +102,33 @@ export class IWantTracer {
   /**
    * Someone delivered a message, stop tracking promises for it
    */
-  async deliverMessage(msgIdStr: string): Promise<void> {
-    this.promises.delete(msgIdStr)
+  deliverMessage(msgIdStr: MsgIdStr): void {
+    this.trackMessage(msgIdStr)
+
+    const expireByPeer = this.promises.get(msgIdStr)
+
+    // Expired promise, check requestMsByMsg
+    if (expireByPeer) {
+      this.promises.delete(msgIdStr)
+
+      if (this.metrics) {
+        this.metrics.iwantPromiseResolved.inc(1)
+        this.metrics.iwantPromiseResolvedPeers.inc(expireByPeer.size)
+      }
+    }
   }
 
   /**
    * A message got rejected, so we can stop tracking promises and let the score penalty apply from invalid message delivery,
    * unless its an obviously invalid message.
    */
-  async rejectMessage(msgIdStr: string, reason: string): Promise<void> {
+  rejectMessage(msgIdStr: MsgIdStr, reason: RejectReason): void {
+    this.trackMessage(msgIdStr)
+
+    // A message got rejected, so we can stop tracking promises and let the score penalty apply.
+    // With the expection of obvious invalid messages
     switch (reason) {
-      case ERR_INVALID_SIGNATURE:
-      case ERR_MISSING_SIGNATURE:
+      case RejectReason.Error:
         return
     }
 
@@ -88,5 +137,30 @@ export class IWantTracer {
 
   clear(): void {
     this.promises.clear()
+  }
+
+  prune(): void {
+    const maxMs = Date.now() - this.requestMsByMsgExpire
+
+    for (const [k, v] of this.requestMsByMsg.entries()) {
+      if (v < maxMs) {
+        // messages that stay too long in the requestMsByMsg map, delete
+        this.requestMsByMsg.delete(k)
+      } else {
+        // recent messages, keep them
+        // sort by insertion order
+        break
+      }
+    }
+  }
+
+  private trackMessage(msgIdStr: MsgIdStr): void {
+    if (this.metrics) {
+      const requestMs = this.requestMsByMsg.get(msgIdStr)
+      if (requestMs !== undefined) {
+        this.metrics.iwantPromiseDeliveryTime.observe((Date.now() - requestMs) / 1000)
+        this.requestMsByMsg.delete(msgIdStr)
+      }
+    }
   }
 }
